Major LocationService Back-end update
Cleaned up a bit of the code already present in the application with
naming of classes.

Added a location service that starts a thread, that does the heavy
lifting. It sends local broadcasts which can be picked up by the
receiver. ATM the receiver doesn't do anything with the messages it
receives from the client, other than to log them, but this can be easily
expanded.