closest distance determination update
At the moment, the input for the spinner is static. However, once I know
how to contact the server for get requests on the place names/locations
available, these will be stored in the current shell class called
TannoyZones. For the moment, the closest location is indicated in the
spinner by a * because bolding the text is more complicated AFAIK.