Minor changes bug fixes
Some receivers were being duplicated unintentionally. Didn't result in
any thing wrong happening, just wasted memory. Also, I realised that
using the startActivity(a) code doesn't let the main activity remember
its state. The easiest solution was to treat the exits to the login
screen and announcement screen as back button presses. I've tested this
and the fix works. I was just concerned it might be something some of
the audience notice and grill us on.